+++ app/src/main/AndroidManifest.xml
@@ -58,7 +58,7 @@
                 <category android:name="android.intent.category.DEFAULT" />
 
                 <data
-                    android:host="your_deeplink_hostname.page.link"
+                    android:host="rocketchat.page.link"
                     android:scheme="https" />
 
             </intent-filter>
