+++ app/src/main/java/chat/rocket/android/util/extensions/Intent.kt
@@ -1,0 +1,10 @@
+package chat.rocket.android.util.extensions
+
+import android.app.Activity
+import android.content.Intent
+
+fun Intent.isSupportedLink(activity: Activity): Boolean {
+    return (action == Intent.ACTION_VIEW && data != null &&
+            (data.isDynamicLink(activity) || data.isAuthenticationDeepLink() ||
+                    data.isCustomSchemeRoomLink() || data.isWebSchemeRoomLink()))
+}
