+++ app/src/main/res/values/defaults.xml
@@ -7,7 +7,7 @@
     <string name="play_store_link" translatable="false">https://play.google.com/store/apps/details?id=chat.rocket.android</string>
     <string name="market_link" translatable="false">market://details?id=chat.rocket.android</string>
     <string name="license_url" translatable="false">https://github.com/RocketChat/Rocket.Chat.Android/blob/develop/LICENSE</string>
-    <string name="dynamiclink_host" translatable="false">your_dynamiclink_host.page.link</string>
+    <string name="dynamiclink_host" translatable="false">rocketchat.page.link</string>
     <string name="app_package_name" translatable="false">chat.rocket.android</string>
     <string name="server_url" translatable="false"></string>
 </resources>
